initialized sql queries

+++ onetoone.sql
@@ -4,7 +4,7 @@ CREATE TABLE customers (
         last_name VARCHAR(35));
 CREATE TABLE investors (
         investor_id BIGSERIAL PRIMARY KEY,
-        customer_id INT UNIQUE,
+        customer_id BIGSERIAL UNIQUE,
         profile_data VARCHAR(256),
         amount NUMERIC(16,2),
             FOREIGN KEY (customer_id) REFERENCES customers(customer_id)
